DB関連のファイルを分離
