+++ python/workdir/connectDB/CreateDB.py
@@ -1,5 +1,7 @@
-from MyDatabase import my_open, my_query, my_close
+from python.workdir.connectDB.MyDatabase import my_open, my_query, my_close
 import mysql.connector
+from DButil import create_database
+
 
 # Data Source Nameのパラメータを辞書型変数で定義
 dsn = {
@@ -9,32 +11,6 @@ dsn = {
     'password': '1234',    # ユーザidに対応するパスワード
     'database': 'sampledb' # オープンするデータベース名
 }
-
-
-def create_database(database_name: str):
-    """
-    データベースの作成
-    Args:
-        database_name (str): 作成するデータベースの名前
-    """
-    dsn_for_creation = dsn.copy()
-    if 'database' in dsn_for_creation:
-        del dsn_for_creation['database']
-    
-    dbcon = None
-    cur = None
-    try:
-        dbcon = mysql.connector.connect(**dsn_for_creation)
-        cur = dbcon.cursor()
-        cur.execute(f"CREATE DATABASE IF NOT EXISTS {database_name} CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci")
-        print(f"Database '{database_name}' created or already exists.")
-    except mysql.connector.Error as err:
-        print(f"Failed to create database: {err}")
-    finally:
-        if cur:
-            cur.close()
-        if dbcon:
-            dbcon.close()
 
 
 def create_tables():
